ps2sdk compiler warning fixes and pfsDopen wrapper Contributed by ragnarok2040 http://forums.ps2dev.org/viewtopic.php?t=11503 Changelog: I initialized some return variables in stdio.c. I added a pfsDopen wrapper for pfsOpen in iop/hdd/pfs/src/pfs_fio.c to fix an incompatible pointer warning for iomanx. I redefined the erl_loader_t type in the ee/erl code and some function parameters to remove the warnings about losing the const qualifier. I added some -fno-builtin-(symbol) flags to EE_LDFLAGS to the ee/erl-loader/Makefile since all of the symbols used aren't builtin versions. I also commented out the full line of rm $(EE_SRC_DIR)exports.c so it doesn't appear during make. I removed the dereference of the return of sio_gets in iop/debug/sior/src/sior.c. I casted ret to (char*) in ee/rpc/sior/src/sior_rpc.c for the return of sio_gets. I converted the imports.lst of iop/tcpip/tcpips to win32 line endings to remove the 'no newline at end of file' warning. I casted the char subscripts in iop/sound/ahx/src/AHX.c to int. I added the declaration of an external fptosi symbol in ee/math/src/sinf.c

+++ iop/sound/ahx/src/AHX.c
@@ -1086,8 +1086,8 @@ void AHXOutput_MixChunk(int NrSamples, int** mb)
 			if(Oversampling) {
 				for(i = 0; i < thiscount; i++) {
 					offset = pos[v] >> 16;
-					sample1 = VolTab[Voices[v].VoiceBuffer[offset]];
-					sample2 = VolTab[Voices[v].VoiceBuffer[offset+1]];
+					sample1 = VolTab[(int)Voices[v].VoiceBuffer[offset]];
+					sample2 = VolTab[(int)Voices[v].VoiceBuffer[offset+1]];
 					frac1 = pos[v] & ((1 << 16) - 1);
 					frac2 = (1 << 16) - frac1;
 					(*mb)[mixpos++] += ((sample1 * frac2) + (sample2 * frac1)) >> 16;
@@ -1095,7 +1095,7 @@ void AHXOutput_MixChunk(int NrSamples, int** mb)
 				}
 			} else {
 				for(i = 0; i < thiscount; i++) {
-					(*mb)[mixpos++] += VolTab[Voices[v].VoiceBuffer[pos[v] >> 16]];
+					(*mb)[mixpos++] += VolTab[(int)Voices[v].VoiceBuffer[pos[v] >> 16]];
 					pos[v] += delta;
 				}
 			}
